database reset related changes and window titles have been changed

+++ estorecpp/src/essystemsettings.cpp
@@ -1,5 +1,6 @@
 #include "essystemsettings.h"
 #include "QSqlQuery"
+#include "utility\utility.h"
 
 ESSystemSettings::ESSystemSettings(QWidget *parent /*= 0*/)
 : QWidget(parent)
@@ -22,10 +23,18 @@ void ESSystemSettings::resetDatabase()
 	{
 	case 0:
 	{
-		QSqlQuery q1("DELETE FROM bill");
-		QSqlQuery q2("DELETE FROM sale");
-		QSqlQuery q3("ALTER TABLE bill AUTO_INCREMENT = 1");
-		QSqlQuery q4("ALTER TABLE sale AUTO_INCREMENT = 1");
+			  if (ES::Utility::verifyUsingMessageBox(this, "Progex", "Do you really want to reset the database this?"))
+			  {
+				  QSqlQuery q1("DELETE FROM bill");
+				  QSqlQuery q2("DELETE FROM sale");
+				  QSqlQuery q3("DELETE FROM return_item");
+				  QSqlQuery q4("DELETE FROM return_bill");
+				  QSqlQuery q5("ALTER TABLE bill AUTO_INCREMENT = 1");
+				  QSqlQuery q6("ALTER TABLE sale AUTO_INCREMENT = 1");
+				  QSqlQuery q7("ALTER TABLE return_item AUTO_INCREMENT = 1");
+				  QSqlQuery q8("ALTER TABLE return_bill AUTO_INCREMENT = 1");
+				  this->close();
+			  }
 		break;
 	}
 	default:
